# Fast computation of Cross Correlation metric