--- a/dipy/align/crosscorr.pyx
+++ b/dipy/align/crosscorr.pyx
@@ -19,6 +19,7 @@ cdef inline int _int_min(int a, int b) nogil:
     """
     return a if a <= b else b
 
+
 cdef enum:
     SI = 0
     SI2 = 1
@@ -27,17 +28,115 @@ cdef enum:
     SIJ = 4
     CNT = 5
 
+
 @cython.boundscheck(False)
 @cython.wraparound(False)
 @cython.cdivision(True)
-def precompute_cc_factors_3d(floating[:, :, :] static, floating[:, :, :] moving,
-                             cnp.npy_intp radius):
+cdef inline int _wrap(int x, int m)nogil:
+    r""" Auxiliary function to `wrap` an array around its low-end side
+    Negative indices are mapped to last coordinates so that no extra memory
+    is required to account for local rectangular windows that exceed the
+    array's low-end boundary.
+
+    Parameters
+    ----------
+    x : int
+        the array position to be wrapped
+    m : int
+        array length
+    """
+    if x<0:
+        return x + m
+    return x
+
+
+@cython.boundscheck(False)
+@cython.wraparound(False)
+@cython.cdivision(True)
+cdef inline void _update_factors(double[:,:,:,:] factors,
+                                 floating[:,:,:] moving,
+                                 floating[:,:,:] static,
+                                 int ss, int rr, int cc,
+                                 int s, int r, int c, int operation)nogil:
+    r"""Updates the precomputed CC factors of a rectangular window
+
+    Updates the precomputed CC factors of the rectangular window centered
+    at (`ss`, `rr`, `cc`) by adding the factors corresponding to voxel
+    (`s`, `r`, `c`) of input images `moving` and `static`.
+
+    Parameters
+    ----------
+    factors : array, shape (S, R, C, 5)
+        array containing the current precomputed factors to be updated
+    moving : array, shape (S, R, C)
+        the moving volume (notice that both images must already be in a common
+        reference domain, i.e. the same S, R, C)
+    static : array, shape (S, R, C)
+        the static volume, which also defines the reference registration domain
+    ss : int
+        first coordinate of the rectangular window to be updated
+    rr : int
+        second coordinate of the rectangular window to be updated
+    cc : int
+        third coordinate of the rectangular window to be updated
+    s: int
+        first coordinate of the voxel the local window should be updated with
+    r: int
+        second coordinate of the voxel the local window should be updated with
+    c: int
+        third coordinate of the voxel the local window should be updated with
+    operation : int, either -1, 0 or 1
+        indicates whether the factors of voxel (`s`, `r`, `c`) should be
+        added to (`operation`=1), subtracted from (`operation`=-1), or set as
+        (`operation`=0) the current factors for the rectangular window centered
+        at (`ss`, `rr`, `cc`).
+
+    """
+    cdef:
+        double sval
+        double mval
+    if s>=moving.shape[0] or r>=moving.shape[1] or c>=moving.shape[2]:
+        if operation == 0:
+            factors[ss, rr, cc, SI] = 0
+            factors[ss, rr, cc, SI2] = 0
+            factors[ss, rr, cc, SJ] = 0
+            factors[ss, rr, cc, SJ2] = 0
+            factors[ss, rr, cc, SIJ] = 0
+    else:
+        sval = static[s,r,c]
+        mval = moving[s,r,c]
+        if operation == 0:
+            factors[ss, rr, cc, SI] = sval
+            factors[ss, rr, cc, SI2] = sval*sval
+            factors[ss, rr, cc, SJ] = mval
+            factors[ss, rr, cc, SJ2] = mval*mval
+            factors[ss, rr, cc, SIJ] = sval*mval
+        elif operation == -1:
+            factors[ss, rr, cc, SI] -= sval
+            factors[ss, rr, cc, SI2] -= sval*sval
+            factors[ss, rr, cc, SJ] -= mval
+            factors[ss, rr, cc, SJ2] -= mval*mval
+            factors[ss, rr, cc, SIJ] -= sval*mval
+        elif operation == 1:
+            factors[ss, rr, cc, SI] += sval
+            factors[ss, rr, cc, SI2] += sval*sval
+            factors[ss, rr, cc, SJ] += mval
+            factors[ss, rr, cc, SJ2] += mval*mval
+            factors[ss, rr, cc, SIJ] += sval*mval
+
+
+@cython.boundscheck(False)
+@cython.wraparound(False)
+@cython.cdivision(True)
+def precompute_cc_factors_3d(floating[:, :, :] static,
+                             floating[:, :, :] moving,
+                             cnp.npy_intp radius, num_threads=None):
     r"""Precomputations to quickly compute the gradient of the CC Metric
 
     Pre-computes the separate terms of the cross correlation metric and image
     norms at each voxel considering a neighborhood of the given radius to
     efficiently compute the gradient of the metric with respect to the
-    deformation field [Avants08][Avants11]
+    deformation field [Ocegueda16][Avants08][Avants11].
 
     Parameters
     ----------
@@ -61,6 +160,10 @@ def precompute_cc_factors_3d(floating[:, :, :] static, floating[:, :, :] moving,
 
     References
     ----------
+    [Ocegueda16] Ocegueda, O., Dalmau, O., Garyfallidis, E., Descoteaux, M.,
+               & Rivera, M. (2016), On the computation of integrals over
+               fixed-size rectangles of arbitrary dimension, Pattern
+               Recognition Letters
     [Avants08] Avants, B. B., Epstein, C. L., Grossman, M., & Gee, J. C. (2008)
                Symmetric Diffeomorphic Image Registration with
                Cross-Correlation: Evaluating Automated Labeling of Elderly and
@@ -69,83 +172,122 @@ def precompute_cc_factors_3d(floating[:, :, :] static, floating[:, :, :] moving,
                Advanced Normalization Tools ( ANTS ), 1-35.
     """
     cdef:
-        cnp.npy_intp side = 2 * radius + 1
         cnp.npy_intp ns = static.shape[0]
         cnp.npy_intp nr = static.shape[1]
         cnp.npy_intp nc = static.shape[2]
-        cnp.npy_intp s, r, c, k, i, j, t, q, qq, firstc, lastc, firstr, lastr
-        double Imean, Jmean
+        cnp.npy_intp side = 2 * radius + 1
+        cnp.npy_intp firstc, lastc, firstr, lastr, firsts, lasts
+        cnp.npy_intp s, r, c, it, sides, sider, sidec
+        double cnt
+        cnp.npy_intp ssss, sss, ss, rr, cc, prev_ss, prev_rr, prev_cc
+        double Imean, Jmean, IJprods, Isq, Jsq
+        double[:, :, :, :] temp = np.zeros((2, nr, nc, 5), dtype=np.float64)
         floating[:, :, :, :] factors = np.zeros((ns, nr, nc, 5),
                                                 dtype=np.asarray(static).dtype)
-        double[:, :] lines = np.zeros((6, side), dtype=np.float64)
-        double[:] sums = np.zeros((6,), dtype=np.float64)
 
     with nogil:
-        for r in range(nr):
-            firstr = _int_max(0, r - radius)
-            lastr = _int_min(nr - 1, r + radius)
-            for c in range(nc):
-                firstc = _int_max(0, c - radius)
-                lastc = _int_min(nc - 1, c + radius)
-                # compute factors for line [:,r,c]
-                for t in range(6):
-                    for q in range(side):
-                        lines[t,q] = 0
+        sss = 1
+        for s in range(ns+radius):
+            ss = _wrap(s - radius, ns)
+            sss = 1 - sss
+            firsts = _int_max(0, ss - radius)
+            lasts = _int_min(ns - 1, ss + radius)
+            sides = (lasts - firsts + 1)
+            for r in range(nr+radius):
+                rr = _wrap(r - radius, nr)
+                firstr = _int_max(0, rr - radius)
+                lastr = _int_min(nr - 1, rr + radius)
+                sider = (lastr - firstr + 1)
+                for c in range(nc+radius):
+                    cc = _wrap(c - radius, nc)
+                    # New corner
+                    _update_factors(temp, moving, static,
+                                       sss, rr, cc, s, r, c, 0)
 
-                # Compute all slices and set the sums on the fly
-                # compute each slice [k, i={r-radius..r+radius}, j={c-radius,
-                # c+radius}]
-                for k in range(ns):
-                    q = k % side
-                    for t in range(6):
-                        sums[t] -= lines[t, q]
-                        lines[t, q] = 0
-                    for i in range(firstr, lastr + 1):
-                        for j in range(firstc, lastc + 1):
-                            lines[SI, q] += static[k, i, j]
-                            lines[SI2, q] += static[k, i, j] * static[k, i, j]
-                            lines[SJ, q] += moving[k, i, j]
-                            lines[SJ2, q] += moving[k, i, j] * moving[k, i, j]
-                            lines[SIJ, q] += static[k, i, j] * moving[k, i, j]
-                            lines[CNT, q] += 1
+                    # Add signed sub-volumes
+                    if s>0:
+                        prev_ss = 1 - sss
+                        for it in range(5):
+                            temp[sss, rr, cc, it] += temp[prev_ss, rr, cc, it]
+                        if r>0:
+                            prev_rr = _wrap(rr-1, nr)
+                            for it in range(5):
+                                temp[sss, rr, cc, it] -= \
+                                    temp[prev_ss, prev_rr, cc, it]
+                            if c>0:
+                                prev_cc = _wrap(cc-1, nc)
+                                for it in range(5):
+                                    temp[sss, rr, cc, it] += \
+                                        temp[prev_ss, prev_rr, prev_cc, it]
+                        if c>0:
+                            prev_cc = _wrap(cc-1, nc)
+                            for it in range(5):
+                                temp[sss, rr, cc, it] -= \
+                                    temp[prev_ss, rr, prev_cc, it]
+                    if(r>0):
+                        prev_rr = _wrap(rr-1, nr)
+                        for it in range(5):
+                            temp[sss, rr, cc, it] += \
+                                temp[sss, prev_rr, cc, it]
+                        if(c>0):
+                            prev_cc = _wrap(cc-1, nc)
+                            for it in range(5):
+                                temp[sss, rr, cc, it] -= \
+                                    temp[sss, prev_rr, prev_cc, it]
+                    if(c>0):
+                        prev_cc = _wrap(cc-1, nc)
+                        for it in range(5):
+                            temp[sss, rr, cc, it] += temp[sss, rr, prev_cc, it]
 
-                    for t in range(6):
-                        sums[t] = 0
-                        for qq in range(side):
-                            sums[t] += lines[t, qq]
-                    if(k >= radius):
-                        # s is the voxel that is affected by the cube with
-                        # slices [s - radius..s + radius, :, :]
-                        s = k - radius
-                        Imean = sums[SI] / sums[CNT]
-                        Jmean = sums[SJ] / sums[CNT]
-                        factors[s, r, c, 0] = static[s, r, c] - Imean
-                        factors[s, r, c, 1] = moving[s, r, c] - Jmean
-                        factors[s, r, c, 2] = (sums[SIJ] - Jmean * sums[SI] -
-                            Imean * sums[SJ] + sums[CNT] * Jmean * Imean)
-                        factors[s, r, c, 3] = (sums[SI2] - Imean * sums[SI] -
-                            Imean * sums[SI] + sums[CNT] * Imean * Imean)
-                        factors[s, r, c, 4] = (sums[SJ2] - Jmean * sums[SJ] -
-                            Jmean * sums[SJ] + sums[CNT] * Jmean * Jmean)
-                # Finally set the values at the end of the line
-                for s in range(ns - radius, ns):
-                    # this would be the last slice to be processed for voxel
-                    # [s, r, c], if it existed
-                    k = s + radius
-                    q = k % side
-                    for t in range(6):
-                        sums[t] -= lines[t, q]
-                    Imean = sums[SI] / sums[CNT]
-                    Jmean = sums[SJ] / sums[CNT]
-                    factors[s, r, c, 0] = static[s, r, c] - Imean
-                    factors[s, r, c, 1] = moving[s, r, c] - Jmean
-                    factors[s, r, c, 2] = (sums[SIJ] - Jmean * sums[SI] -
-                        Imean * sums[SJ] + sums[CNT] * Jmean * Imean)
-                    factors[s, r, c, 3] = (sums[SI2] - Imean * sums[SI] -
-                        Imean * sums[SI] + sums[CNT] * Imean * Imean)
-                    factors[s, r, c, 4] = (sums[SJ2] - Jmean * sums[SJ] -
-                        Jmean * sums[SJ] + sums[CNT] * Jmean * Jmean)
-    return np.asarray(factors)
+                    # Add signed corners
+                    if s>=side:
+                        _update_factors(temp, moving, static,
+                            sss, rr, cc, s-side, r, c, -1)
+                        if r>=side:
+                            _update_factors(temp, moving, static,
+                                sss, rr, cc, s-side,r-side, c, 1)
+                            if c>=side:
+                                _update_factors(temp, moving, static,
+                                    sss, rr, cc, s-side, r-side, c-side, -1)
+                        if c>=side:
+                            _update_factors(temp, moving, static,
+                                sss, rr, cc, s-side, r, c-side, 1)
+                    if r>=side:
+                        _update_factors(temp, moving, static,
+                            sss, rr, cc, s, r-side, c, -1)
+                        if c>=side:
+                            _update_factors(temp, moving, static,
+                                sss, rr, cc, s, r-side, c-side, 1)
+
+                    if c>=side:
+                        _update_factors(temp, moving, static,
+                            sss, rr, cc, s, r, c-side, -1)
+                    # Compute final factors
+                    if s>=radius and r>=radius and c>=radius:
+                        firstc = _int_max(0, cc - radius)
+                        lastc = _int_min(nc - 1, cc + radius)
+                        sidec = (lastc - firstc + 1)
+                        cnt = sides*sider*sidec
+                        Imean = temp[sss, rr, cc, SI] / cnt
+                        Jmean = temp[sss, rr, cc, SJ] / cnt
+                        IJprods = (temp[sss, rr, cc, SIJ] -
+                            Jmean * temp[sss, rr, cc, SI] -
+                            Imean * temp[sss, rr, cc, SJ] +
+                            cnt * Jmean * Imean)
+                        Isq = (temp[sss, rr, cc, SI2] -
+                            Imean * temp[sss, rr, cc, SI] -
+                            Imean * temp[sss, rr, cc, SI] +
+                            cnt * Imean * Imean)
+                        Jsq = (temp[sss, rr, cc, SJ2] -
+                            Jmean * temp[sss, rr, cc, SJ] -
+                            Jmean * temp[sss, rr, cc, SJ] +
+                            cnt * Jmean * Jmean)
+                        factors[ss, rr, cc, 0] = static[ss, rr, cc] - Imean
+                        factors[ss, rr, cc, 1] = moving[ss, rr, cc] - Jmean
+                        factors[ss, rr, cc, 2] = IJprods
+                        factors[ss, rr, cc, 3] = Isq
+                        factors[ss, rr, cc, 4] = Jsq
+    return factors
 
 
 @cython.boundscheck(False)