This PR implements a new algorithm for computing the Cross Correlation metric, which is substantially faster than the original algorithm implemented in ANTS. The computational complexity of the original algorithm for a volume of size N^3 using rectangular windows of size K^3 is Theta(N^3 * K^2). This means that the time increases quadratically with the window radius. 

The computational complexity of this new algorithm is Theta(N^3), it does not depend on the window size, so we observe a substantial speed up, especially for large rectangular windows. After a conversation with @grlee77, the plan would be to add this algorithm instead of #1051, and then after #1050 is in, we can go ahead and use that framework to parallelize this sequential algorithm to obtain even higher speed up.

These are some results on a 4-Core i5-4200H CPU @ 2.80GHz:
Old algorithm, parallel (using #1051):
1 thread
  pre: 1.23728 s
2 threads
  pre: 0.72080 s
4 threads (default)
  pre: 0.69974 s

New algorithm (this PR)
1 thread
  pre: 0.15835 s
2 threads
  pre: 0.15837 s
4 threads (default)
  pre: 0.15925 s
(of course it doesn't depend on the number of threads, it's not parallel right now)

Reference: http://www.sciencedirect.com/science/article/pii/S0167865516300861