@grlee77 2016-06-02
The code for the new algorithm looks good and the performance is great.  I also found the new algorithm proposed here to be faster than the multithreaded version of the old algorithm (in #1051). 

Unlike the previous algorithm, this one is not trivial to make parallel, though.  The main difficulty is that the computations for a given slice, row, or column also make use of previously computed results from the previous slice, row and/or column. If we split slices across threads for instance then we can no longer assume slices are computed sequentially which seems to be required here.  There may be some clever way around this, but it is no longer a trivial case of just using `prange` for the outer loop.