@omarocegueda 2016-06-03
Thanks for testing the algorithm @grlee77 !
I just finished a prototype of how to implement this algorithm in parallel. We just assign consecutive slices to each thread. You are right, this algorithm references the previous sums but we can solve this issue by allowing the slice intervals to overlap: at the beginning, each thread must compute the result for the first slice using the direct algorithm. For a volume of NxNxN, computing integrals over windows of KxKxK using T threads, we will be making O(T*K*N^2) total extra work (i.e. T*K extra slices because of the overlap), which will be effective if the window size, K, is not too big and we use just a few cores T.

Here is the kind of performance I am observing:

**First verify that both approaches are equivalent:**
Maximum difference (direct vs. fast-sequential): 0.000000
Maximum difference (direct vs. fast-parallel[1 threads]): 0.000000
Maximum difference (direct vs. fast-parallel[2 threads]): 0.000000
Maximum difference (direct vs. fast-parallel[3 threads]): 0.000000
Maximum difference (direct vs. fast-parallel[4 threads]): 0.000000

**Now check performance (on a computer with 2 physical cores i5-4200H  @ 2.80GHz):**
Fast sequential. Elapsed: 0.352472
1 threads. Elapsed: 0.364174<<<< A bit slower than the sequential algorithm
2 threads. Elapsed: 0.196001<<<< We only have 2 physical cores
3 threads. Elapsed: 0.250694
4 threads. Elapsed: 0.200792

We can see that it is using the 2 cores very efficiently. You can find this proof of concept here:
https://github.com/omarocegueda/fast_integrals_parallel

We can implement this strategy here once your PR is merged =)