@omarocegueda 2016-09-30
Thank you very much for your thorough review Ariel!, =D