@arokem 2016-09-30
Thank you for the quick turnaround! I will merge this early next week, unless some else pipes up.